rename brushes
