+++ index.ts
@@ -337,11 +337,11 @@ import { Control } from './src/controls/Control';
 import { Gradient } from './src/gradient/Gradient';
 import { Pattern } from './src/Pattern';
 import { Shadow } from './src/Shadow';
-import { BaseBrush } from './src/brushes/base_brush.class';
-import { PencilBrush } from './src/brushes/pencil_brush.class';
-import { CircleBrush } from './src/brushes/circle_brush.class';
-import { SprayBrush } from './src/brushes/spray_brush.class';
-import { PatternBrush } from './src/brushes/pattern_brush.class';
+import { BaseBrush } from './src/brushes/BaseBrush';
+import { PencilBrush } from './src/brushes/PencilBrush';
+import { CircleBrush } from './src/brushes/CircleBrush';
+import { SprayBrush } from './src/brushes/SprayBrush';
+import { PatternBrush } from './src/brushes/PatternBrush';
 import { FabricObject as Object } from './src/shapes/Object/FabricObject';
 import { Line } from './src/shapes/Line';
 import { Circle } from './src/shapes/Circle';
