+++ src/brushes/BaseBrush.ts
