+++ src/brushes/PatternBrush.ts
@@ -2,7 +2,7 @@ import { Pattern } from '../Pattern';
 import { PathData } from '../typedefs';
 import { createCanvasElement } from '../util/misc/dom';
 import type { Canvas } from '../canvas/CanvasEvents';
-import { PencilBrush } from './pencil_brush.class';
+import { PencilBrush } from './PencilBrush';
 
 export class PatternBrush extends PencilBrush {
   declare source?: CanvasImageSource;
